+++ greenmap/src/main/java/com/ucamp/greenmap/point/domain/PointHistory.java
@@ -30,7 +30,7 @@ public class PointHistory extends BaseEntity {
     private Category category;
 
     @Column(name = "point_amount", nullable = false)
-    private Integer pointAmount;
+    private Long pointAmount;
 
     @Column(name = "description")
     private String description;
